requirements.txt
asgiref==3.8.1
Django==5.1.1
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
